flask_server/requirements.txt
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
dlib==19.24.2
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==3.0.0
idna==3.5
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.26.2
Pillow==10.1.0
requests==2.31.0
urllib3==2.1.0
Werkzeug==3.0.1
